boom

+++ app/Http/Controllers/MainController.php
@@ -72,10 +72,14 @@ class MainController extends Controller
 
     public function selectedDelete(Request $request)
     {
-        dd(1);
-        $studId = $request->issetStud;
-        CardStudent::where('id', $studId)->delete();
-        //return redirect()->back();
+        if( $request->ajax() ) {
+
+            $studId = $request->response;
+            CardStudent::where('id', $studId)->delete();
+        }
+        
+        
+        return $request->response;
     }
 
     public function searchPost()
